Country Guide App
In this tutotial, I'm going to create a Country guide app or Display country information app. This app allows the user to search for a particular country and then displays information related to the searched country. For this, I will use "Rest Country API"
When user will write the country name in the input field, then the country name will be suggested automatically. To do this, I will also use 'Rest Countries API'.
I hope, It will be the best API projects for the developers who want to get data from anywhere in the web.